src/main/requirements.txt
PyGitHub==2.3.0
python-dotenv==1.0.1
Jinja2==3.1.4
python-gitlab==4.6.0
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.1.0
black==24.4.2
